Performance - fix re-renders
